requirements.txt
# Core dependencies
aiohttp>=3.8.0
asyncio
python-dotenv>=0.19.0

# Solana and blockchain
solana>=0.30.0
solders>=0.18.0
anchorpy>=0.18.0
base58>=2.1.1

# Data handling
pandas>=1.5.0
numpy>=1.24.0

# Logging and utilities
colorlog>=6.7.0
python-dateutil>=2.8.0

# API and HTTP
requests>=2.28.0
urllib3>=1.26.0
openai>=1.3.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Optional: Database support
sqlalchemy>=1.4.0

# Optional: Advanced analytics
scipy>=1.10.0
scikit-learn>=1.2.0 